[Editor/ModManager/ModSupport] Prevent crashes | Commit by Wannkunstbeikor | https://github.com/CadeEvs/FrostyToolsuite/pull/242
- Prevents crash when asset is modified by handler and non handler mod
- Prevents crash when loading default profile
- Prevents crash when frosty fails to decompress archive
- Replaces symlink exception with message box

+++ FrostyModSupport/FrostyModExecutor.cs
@@ -470,6 +470,11 @@ namespace Frosty.ModSupport
                                 if (existingEntry.Sha1 == resource.Sha1)
                                     return;
 
+                                if (!m_archiveData.ContainsKey(existingEntry.Sha1))
+                                {
+                                    return;
+                                }
+
                                 m_archiveData[existingEntry.Sha1].RefCount--;
                                 if (m_archiveData[existingEntry.Sha1].RefCount == 0)
                                     m_archiveData.TryRemove(existingEntry.Sha1, out _);
@@ -563,6 +568,11 @@ namespace Frosty.ModSupport
                                     return;
                                 if (existingEntry.Sha1 == resource.Sha1)
                                     return;
+
+                                if (!m_archiveData.ContainsKey(existingEntry.Sha1))
+                                {
+                                    return;
+                                }
 
                                 m_archiveData[existingEntry.Sha1].RefCount--;
                                 if (m_archiveData[existingEntry.Sha1].RefCount == 0)
@@ -668,6 +678,11 @@ namespace Frosty.ModSupport
                                 ChunkAssetEntry existingEntry = m_modifiedChunks[guid];
                                 if (existingEntry.Sha1 == resource.Sha1)
                                     return;
+
+                                if (!m_archiveData.ContainsKey(existingEntry.Sha1))
+                                {
+                                    return;
+                                }
 
                                 m_archiveData[existingEntry.Sha1].RefCount--;
                                 if (m_archiveData[existingEntry.Sha1].RefCount == 0)
@@ -971,6 +986,11 @@ namespace Frosty.ModSupport
                         if (existingEntry.Sha1 == resource.GetValue<Sha1>("sha1"))
                             continue;
 
+                        if (!m_archiveData.ContainsKey(existingEntry.Sha1))
+                        {
+                            return;
+                        }
+
                         m_archiveData[existingEntry.Sha1].RefCount--;
                         if (m_archiveData[existingEntry.Sha1].RefCount == 0)
                             m_archiveData.TryRemove(existingEntry.Sha1, out _);
@@ -1022,6 +1042,11 @@ namespace Frosty.ModSupport
                         ResAssetEntry existingEntry = m_modifiedRes[name];
                         if (existingEntry.Sha1 == resource.GetValue<Sha1>("sha1"))
                             continue;
+
+                        if (!m_archiveData.ContainsKey(existingEntry.Sha1))
+                        {
+                            return;
+                        }
 
                         m_archiveData[existingEntry.Sha1].RefCount--;
                         if (m_archiveData[existingEntry.Sha1].RefCount == 0)
@@ -1076,6 +1101,11 @@ namespace Frosty.ModSupport
                         ChunkAssetEntry existingEntry = m_modifiedChunks[chunkId];
                         if (existingEntry.Sha1 == resource.GetValue<Sha1>("sha1"))
                             continue;
+
+                        if (!m_archiveData.ContainsKey(existingEntry.Sha1))
+                        {
+                            return;
+                        }
 
                         m_archiveData[existingEntry.Sha1].RefCount--;
                         if (m_archiveData[existingEntry.Sha1].RefCount == 0)
@@ -1536,8 +1566,13 @@ namespace Frosty.ModSupport
                     FrostyMessageBox.Show(reason + "\r\n\r\nShortly you will be prompted for elevated privileges, this is required to create symbolic links between the original data and the new modified data. Please ensure that you accept this to avoid any issues.", "Frosty Toolsuite");
                     if (!RunSymbolicLinkProcess(cmdArgs))
                     {
-                        Directory.Delete(modDataPath, true);
-                        throw new FrostySymLinkException();
+                        FrostyMessageBox.Show("Frosty needs to generate symbolic links, please ensure that you accept this so you don't have to regenerate ModData.", "Frosty Editor");
+                        if (!RunSymbolicLinkProcess(cmdArgs))
+                        {
+                            Directory.Delete(modDataPath, true);
+                            FrostyMessageBox.Show("One ore more symbolic links could not be created, please restart tool as Administrator and ensure your storage drive is formatted to NTFS (not exFAT).", "Frosty Editor");
+                            return -1;
+                        }
                     }
                 }
 
@@ -2715,7 +2750,10 @@ namespace Frosty.ModSupport
             ExecuteProcess("cmd.exe", "/C \"" + AppDomain.CurrentDomain.BaseDirectory + "\\run.bat\"", true, true);
 
             // delete batch
-            File.Delete("run.bat");
+            if (File.Exists("run.bat"))
+            {
+                File.Delete("run.bat");
+            }
 
             // validate
             foreach (SymLinkStruct arg in cmdArgs)
@@ -2752,10 +2790,17 @@ namespace Frosty.ModSupport
                     process.StartInfo.Verb = "runas";
                 }
 
-                process.Start();
+                try
+                {
+                    process.Start();
 
-                if (waitForExit)
-                    process.WaitForExit();
+                    if (waitForExit)
+                        process.WaitForExit();
+                }
+                catch
+                {
+                    // do nothing
+                }
             }
         }
